CORE-16353: Added flow tests using the Flow Testing Driver and fixed a bug on the Accept smart contract
Added flow tests using the Flow Testing Driver and fixed a bug on the Accept smart contract which I identified while executing the newly added flow tests. The bug was allowing proposers to accept their own proposals, for example, Alice issues £20 proposal to Bob, Bob modifies it to £20 and then accepts the modified proposal. Also. modified some of the dependencies to Corda 5.1 as that it what is compatible with the Flow Testing Driver. Obviously this code will not be merged until Corda 5.1 (and the Flow Testing Driver) is available publicly i.e. full GA. Also, updated the README.md file. Also, made changes to the flow classes to return the states' IDs instead of transaction IDs which were not being used anywhere. The returned states' IDs are required in flow tests to be able to test subsequent flows. Also, worth mentioning that I had to change the getter property for IsBuyer as for some weird reason Jackson renames it to just Buyer when serialising it and therefore breaking the flow tests.